requirements.txt
streamlit
streamlit-js-eval
openai
sqlalchemy
azure-cognitiveservices-speech
watchdog
pyodbc
cryptography
